requirements.txt
streamlit
supabase
groq
plotly
pandas
python-dotenv
psycopg2-binary
faker